requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
itsdangerous==0.23
requests==2.1.0
streql==3.0.2
python-dotenv==0.1.2
bigcommerce==0.11.0
